device: add devena-firstboot integration
Due to the diverse nature of bootloader implementations, we can not have
a unified way to generate bootloader configuration automatically. While
this integration automatically installs and creates devena-firstboot
initramfs images, the maintainer has to generate a proper configuration
for the bootloader the device is using, and also configure the
bootloader to load the initramfs image on startup.