context, utils: fill the filesystem free space with zeros
This procedure increases the compression efficiency by wiping any
discarded filesystem blocks with zeros (since the compression algorithms
can deal with these). However, the buffer is way larger than the block
size of a filesystem (1MiB vs 4KiB), but it is fast and good enough.

We are operating on the filesystem (block) level, not on the file level,
and [insert 1000 line of text about  how file deletion works on the
filesystem level here], so that we don't left any discarded but non-zero
blocks behind.